Add routes behavior
