+++ src/app/App.tsx
@@ -1,98 +1,13 @@
-'use client'
+import '../app/pages/start/StartScreen.css';
+import StartScreen from './pages/start/page';
 
-import { useCallback, useEffect, useState } from 'react';
-import StartScreen from './components/StartScreen';
-import Game from './components/Game';
-import GameOver from './components/GameOver';
-import { wordsList } from './data/words';
-import './App.css';
-
-// Definição dos estágios do jogo com tipos explícitos
-const stages = [
-  { id: 1, name: "start" },
-  { id: 2, name: "game" },
-  { id: 3, name: "end" },
-];
-
-interface WordList {
-  [key: string]: string[];
-}
 
 const App = () => {
-  // Utilizando a tipagem explícita para cada estado
-  const [gameStage, setGameStage] = useState<string>(stages[0].name);
-  const [words] = useState<WordList>(wordsList);
-  const [pickedWord, setPickedWord] = useState<string>("");
-  const [pickedCategory, setPickedCategory] = useState<string>("");
-  const [letters, setLetters] = useState<string[]>([]);
-  const [guessedLetters, setGuessedLetters] = useState<string[]>([]);
-  const [wrongLetters, setWrongLetters] = useState<string[]>([]);
-  const [guesses, setGuesses] = useState<number>(3);
-  const [score, setScore] = useState<number>(0);
-
-  const pickedWordAndCategory = useCallback(() => {
-    const categories = Object.keys(words);
-    const category = categories[Math.floor(Math.random() * categories.length)];
-    const word = words[category][Math.floor(Math.random() * words[category].length)];
-    return { word, category };
-  }, [words]);
-
-  const startGame = useCallback(() => {
-    clearLetterStates();
-    const { word, category } = pickedWordAndCategory();
-    let wordLetters = word.split("").map(l => l.toLowerCase());
-    setPickedWord(word);
-    setPickedCategory(category);
-    setLetters(wordLetters);
-    setGameStage(stages[1].name);
-  }, [pickedWordAndCategory]);
-
-  const verifyLetter = (letter: string) => {
-    const normalizedLetter = letter.toLowerCase();
-    if (guessedLetters.includes(normalizedLetter) || wrongLetters.includes(normalizedLetter)) {
-      return;
-    }
-    if (letters.includes(normalizedLetter)) {
-      setGuessedLetters(actualGuessedLetters => [...actualGuessedLetters, normalizedLetter]);
-    } else {
-      setWrongLetters(actualWrongLetters => [...actualWrongLetters, normalizedLetter]);
-      setGuesses(actualGuesses => actualGuesses - 1);
-    }
-  };
-
-  const clearLetterStates = () => {
-    setGuessedLetters([]);
-    setWrongLetters([]);
-  };
-
-  useEffect(() => {
-    if (guesses <= 0) {
-      clearLetterStates();
-      setGameStage(stages[2].name);
-    }
-  }, [guesses]);
-
-  useEffect(() => {
-    const uniqueLetters = Array.from(new Set(letters));
-    if (guessedLetters.length === uniqueLetters.length && uniqueLetters.length > 0) {
-      setScore(actualScore => actualScore + 100);
-      startGame();
-    }
-  }, [guessedLetters, letters, startGame]);
-
-  const retry = () => {
-    setScore(0);
-    setGuesses(3);
-    setGameStage(stages[0].name);
-  };
-
   return (
-    <div className="App">
-      {gameStage === 'start' && <StartScreen startGame={startGame} />}
-      {gameStage === 'game' && <Game verifyLetter={verifyLetter} pickedWord={pickedWord} pickedCategory={pickedCategory} letters={letters} guessedLetters={guessedLetters} wrongLetters={wrongLetters} guesses={guesses} score={score} />}
-      {gameStage === 'end' && <GameOver retry={retry} score={score} />}
-    </div>
+    <>
+      <StartScreen/>
+    </>
   );
-};
+}
 
 export default App;
