--- src/app/components/GameOver.css
@@ -1,5 +1,0 @@
-h2 span {
-    color: yellow;
-    font-size: 1,5em;
-    
-}
