--- src/app/components/StartScreen.css
@@ -1,8 +1,0 @@
-.start h1 {
-    font-size: 3.5em;
-}
-
-.start p {
-    margin-bottom: 2em;
-    color: yellow;
-}
