+++ src/app/data/words.tsx
@@ -3,6 +3,6 @@ export const wordsList = {
     fruta: ["Banana", "Maçã", "Pêra", "Mamão", "Laranja"],
     corpo: ["Braço", "Perna", "Cérebro", "Pescoço", "Olhos"],
     computador: ["Mouse", "Teclado", "Monitor", "Gabinete"],
-    programação: ["Linguagem", "Framework", "JavaScript", "React"],
+    programação: ["Java", "POO", "C#", "React"],
     alimento: ["Arroz", "Feijão", "Carne", "Leite", "Ovo"],
 };
