+++ src/app/pages/gameover/GameOver.css
@@ -1,0 +1,14 @@
+h2 span {
+    color: yellow;
+    font-size: 1,5em;
+    
+}
+
+.container{
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    text-align: center;
+}
