+++ src/app/pages/start/StartScreen.css
@@ -1,0 +1,17 @@
+.start h1 {
+    font-size: 3.5em;
+}
+
+.start p {
+    margin-bottom: 2em;
+    color: yellow;
+}
+
+.start  {
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    text-align: center;
+}
