+++ src/app/pages/start/page.tsx
@@ -1,15 +1,14 @@
 import './StartScreen.css';
+import Link from 'next/link';
 
-interface StartScreenProps {
-  startGame: () => void;
-}
-
-const StartScreen = ({ startGame }: StartScreenProps) => {
+const StartScreen = () => {
   return (
     <div className='start'>
       <h1>Jogo de Palavras</h1>
       <p>Clique no botão abaixo para começar a jogar</p>
-      <button onClick={startGame}>Começar o jogo</button>
+      <Link href={"/pages/game"}>
+        <button>Começar o jogo</button>
+      </Link>
     </div>
   );
 }
